requirements.txt
Flask==3.0.3
jinja2>=3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
flask-login==0.6.3
Pillow>=10.4.0
flask-mail==0.9.1
python-dotenv>=1.2.0
requests
yagmail
